requirements.txt
flask
line-bot-sdk
wheel
grpcio
google-cloud-firestore
flask_cors
line-bot-sdk
google-cloud-storage
gunicorn
google-cloud-logging
firebase_admin
matplotlib
tensorflow==1.13.1
Pillow
numpy
scipy>=1.1.0,<=1.2.1
scikit-image>=0.15.0
h5py>=2.7.1